bus/hp_dio/98265a: convert to anonymous namespace (nw)
